Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Apart from the new style I removed the left and right transparency like how the framework drawables are now formatted in Gingerbread. I don't know if this will make them render incorrectly in the app or not, ie. if the change is something that affects all notification icons or just the ones in the framework. Let me know!

Patch-set: 1