Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Looks good on both MDPI (hero) and HDPI (passion).

attn1: Did you really review the actual code? If so, you should have seen that this feature must be turned on in an overlay, otherwise no change. That's the author's point according to the description!

Patch-set: 1