Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Also, turning this on wont instantly give you seek functionality, is must also be supported by the JNI used on your device. This works for si4708 but I'm not sure if any other chips have the seek function working in JNI. All I was requesting for mdpi was that someone test that the buttons where positioned correctly in the layout.

Since most devices will probably not currently support this the overlay defaults to disabled.

Patch-set: 1