Update patch set 1
Patch Set 1:

Jonas - did you use the mdpi layout in my patch or the one Ladios provided. Which is better. Some of the values in the one Ladios provided don't look right to me.

Burstlam - It searches the correct direction for me. Are you using the updated si4708 code that I uploaded for this?

Patch-set: 1