Update patch set 2
Patch Set 2: No score

So am I correct that getFreq() will return the frequency it is at during the search so if two frequency readings in a row (with a small delay) are the same it indicates the search has finished?
I think it should go into updateSearchProgress(), not the seek methods.
I'll upload a patch a little later.

Patch-set: 2