Update patch set 3
Patch Set 3:

Test that, I think it does what you want.

Is 4 seconds long enough to find a station or should it loop a bit longer.

There is no cancel search jni for si4708, but I assumed the cancelSearch method should be run for other hardware if the loop completes without finding a station.

Patch-set: 3