Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

Uploaded a modified mdpi layout. This time I modified the app so it would run on an mdpi device I have which has not FM hardware.

Looks good to me, someone please test the functionality on an mdpi device with a radio.

Patch-set: 4