Update patch set 4
Patch Set 4:

KalimochoAz - the need to set frequency before seeking is a unique requirement of the wl1271. It should be done as part of the seek function in jni. Forcing all devices to tune a frequency before seeking will needlessly slow the seek process, and it will probably also be quicker for the wl1271 if it is done in one jni call.

Try something like this https://gist.github.com/1029115 (untested. I have no wl1271 hardware)

Patch-set: 4