Update patch set 1
Patch Set 1:

Just a quick update to this. I have been playing with the bcm FM interface on my o2x (audio routing is broken, but I can see what the bcm radio chip is doing). The problem with getFreq is the hci_r function in the jni is broken. I've fixed that, but haven't got seek working properly yet. Once I get seek working I'll submit a patch for it. 

This patch should not be merged (at least not the bit that changes getFreq. The bit that removes all the tabs I added to FMRadio.java looks good).

Patch-set: 1