Update patch set 1
Patch Set 1:

Is there any problems still with the patch referred to in a previous comment? I merged it in my own KANG and it works fine on the N1. I just don't know if it will work on other chips since I believe the 4325 may be different in this from the 4329.

Patch-set: 1