Update patch set 1
Patch Set 1:

If you mean change 6157 I haven't tested it yet, but it looks very similar to what I was going to submit so I expect it will work. It certainly should fix the problem this change was submitted for, the only bit I'm not sure about is the seek functionality. It may need a small change to work correctly on some hardware.

I still think this change should be abandoned and the other changes dependent on this should be resubmitted without the dependency. Merging this when it is known it will break things just so it can be reverted by another change is a bad idea.

Patch-set: 1