Update patch set 1
Patch Set 1:

Do some hardware use a different seek even if it's a BCM chip? Maybe we just need to branch this off to 4325 and 4329 if so since I'm guessing that those two are the only reasons for seek not working the same. Anyways, it needs a thorough test on all BCM devices in order to ensure stability.

Patch-set: 1