Update patch set 1
Patch Set 1:

As I said 6157 should fix the problem this was submitted for (on all bcm devices), its only the seek part that may not work correctly on some hardware. If 6157 is merged as is the seek functionality can be left disabled on any devices that it doesn't function correctly on.

The seek function should be basically the same on all bcm devices, but on the o2x I used a sleep before it returned otherwise it would stop seeking almost immediately and usually not tune far enough to find a station. I'm not sure if all version of the bcm hardware have the same issue. I think when FM app tries to get the newly tuned frequency it was stopping the search, so the jni has to wait long enough for the station to be tuned before returning.

Patch-set: 1