Update patch set 1
Patch Set 1:

but 6157 is the code for that part, so it is also a fix for this issue. 

By the way, this issue has nothing to do with seek, the issue is the function to get the currently tuned frequency on bcm hardware is broken and always returns 64 (because the hci_r function is broken). As well as adding seek functionality 6157 fixes the hci_r function.

I've been busy and haven't done any more testing. I'll try 6157 tonight and see if the sleep is needed. If it is I'll upload a new patchset. I recommend you try 6157, it should fix the problem and also give you seek functionality (but make sure you revert this change or the FM apps display won't update after seek).

Patch-set: 1