Update patch set 2
Patch Set 2:

Ok. I tested without my libaudio patches, and it was the same issue, so none of them should be causing it. It sticks to adjusting ringer volume on both speaker and headset I noticed. If I turn off and on FM, then adjusting media volume works. But if I then turn on speaker, it switched to adjusting ringer volume.

Patch-set: 2