Update patch set 6
Patch Set 6:

Great job.  The only other thing I believe it needs is a call to unMute() somewhere in fmOn().  That will fix the problem of pausing on the lockscreen, turning off the app, turning on the app, and then having the pause button do nothing the first time it is pressed.

Patch-set: 6